Reorganize module structure of engagement_edge
`src` isn't a unique (or appropriate!) top-level module name, and
causes conflicts when trying to type across all our services at the
same time.
