+++ src/python/engagement_edge/BUILD
@@ -1,4 +1,4 @@
 python_awslambda(
     runtime="python3.7",
-    handler="src/engagement_edge.py:app"
+    handler="engagement_edge/engagement_edge.py:app"
 )
