--- src/python/engagement_edge/src/lib/BUILD
@@ -1,1 +1,0 @@
-python_library()
